Added Firebase dependencies
