+++ Chatlin/.idea/misc.xml
@@ -3,14 +3,17 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.205" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/drawable/background_btn.xml" value="0.1" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/drawable/background_btn2.xml" value="0.1" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/drawable/background_btn3.xml" value="0.1" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/drawable/background_edittext.xml" value="0.1" />
-        <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/layout/activity_login.xml" value="0.264" />
+        <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/drawable/ic_launcher_background.xml" value="0.205" />
+        <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/layout/activity_login.xml" value="0.23157894736842105" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/layout/activity_main.xml" value="0.2" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/layout/activity_sign_up.xml" value="0.20521590423257802" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/layout/activity_welcome.xml" value="0.2" />
+        <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/Debugit_2022/Chatlin/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.205" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/deubgit'22/Debugit_2022/Chatlin/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1875" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/deubgit'22/Debugit_2022/Chatlin/app/src/main/res/drawable/background_btn.xml" value="0.12" />
         <entry key="..\:/Users/yashs/OneDrive/document (1)/COPS/deubgit'22/Debugit_2022/Chatlin/app/src/main/res/drawable/background_btn2.xml" value="0.204" />
