+++ Chatlin/app/src/androidTest/java/com/yashacker/chatlin/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.chatlin
+package com.yashacker.chatlin
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
