+++ Chatlin/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.chatlin">
+    package="com.yashacker.chatlin">
 
     <application
         android:allowBackup="true"
@@ -10,23 +10,25 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Chatlin">
         <activity
-            android:name=".login_Activity"
+            android:name="com.yashacker.chatlin.Login_Activity"
             android:exported="false" />
         <activity
-            android:name=".SignUp_Activity"
+            android:name="com.yashacker.chatlin.SignUp_Activity"
             android:exported="false" />
         <activity
-            android:name=".Welcome"
-            android:exported="false" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
+            android:name="com.yashacker.chatlin.Welcome"
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
+            </activity>
+        <activity
+            android:name="com.yashacker.chatlin.MainActivity"
+            android:exported="false"/>
+
+
     </application>
 
 </manifest>
