--- Chatlin/app/src/main/java/com/example/chatlin/Welcome.kt
@@ -1,11 +1,0 @@
-package com.example.chatlin
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class Welcome : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_welcome)
-    }
-}
