+++ Chatlin/app/src/main/res/layout/activity_sign_up.xml
@@ -8,6 +8,7 @@
     tools:context=".SignUp_Activity">
 
     <ImageView
+        android:id="@+id/signupbacktowelcome"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
@@ -63,7 +64,7 @@
             android:textSize="18sp"/>
 
         <EditText
-            android:id="@+id/editTextTextPersonName"
+            android:id="@+id/editUsername"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:ems="10"
@@ -82,7 +83,7 @@
             android:textSize="18sp" />
 
         <EditText
-            android:id="@+id/editTextTextEmailAddress"
+            android:id="@+id/editEmail"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:hint="Your Email"
@@ -102,7 +103,7 @@
             android:textSize="18sp" />
 
         <EditText
-            android:id="@+id/editTextTextPassword"
+            android:id="@+id/editPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
@@ -130,12 +131,12 @@
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:gravity="right"
-            android:text="Don,t have an account?"
+            android:text="Don't have an account?"
             android:textColor="#000000"
             android:textSize="16sp" />
 
         <TextView
-            android:id="@+id/textView8"
+            android:id="@+id/textlogin"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_weight="1"
