+++ Chatlin/build.gradle
@@ -1,10 +1,36 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
+buildscript {
+    ext.kotlin_version= "1.6.20"
+    repositories {
+        google()
+        mavenCentral()
+
+    }
+    dependencies {
+        classpath 'com.google.gms:google-services:4.3.10'
+        classpath "com.android.tools.build:gradle:7.1.3"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+
+    }
+}
+
+
+
 plugins {
     id 'com.android.application' version '7.1.3' apply false
     id 'com.android.library' version '7.1.3' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.20' apply false
 }
 
+allprojects {
+    repositories {
+        google()
+        mavenCentral()
+        jcenter()
+
+
+    }
+}
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
